+++ appseed/openssl/crypto/cms/cms_lib.c
@@ -176,7 +176,7 @@ int CMS_dataFinal(CMS_ContentInfo *cms, BIO *cmsbio)
 		/* Set bio as read only so its content can't be clobbered */
 		BIO_set_flags(mbio, BIO_FLAGS_MEM_RDONLY);
 		BIO_set_mem_eof_return(mbio, 0);
-		ASN1_STRING_set0(*pos, cont, contlen);
+		ASN1_STRING_set0(*pos, cont, (int) contlen);
 		(*pos)->flags &= ~ASN1_STRING_FLAG_CONT;
 		}
